uart: stm32: async api: fix UART_RX_DISABLED event generation
When the current buffer is filled and no valid second buffer is
provided, a UART_RX_DISABLED event must be generated. This is obtained
by forcing the RX timeout to expire soon, so that the receive disable
and DMA stop can be safely called. This change doesn't takes the
circular buffer mode into account. Common code used for all cases when
a UART_RX_BUF_RELEASED event is generated.

Signed-off-by: Giancarlo Stasi <giancarlo.stasi.co@gmail.com>